+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This is used for recording Invader's changes. This changelog is based on
 - invader-dependency: Now accepts filesystem paths for the tag if `-P` is passed
 - invader-font: Now accepts filesystem paths for the the TTF file if `-P` is passed
 - invader-string: Now accepts filesystem paths for the text file if `-P` is passed
+- invader-crc: Now prints a warning to stderr if the CRC32 in the cache file header is wrong
 
 ### Changed
 - Replaced the command line arg parser. The new one was made from scratch and uses only functions from the C++ standard